Terminal font height setting (#2683)
* calculate terminal font like editor

* terminal: calc contrast and switch colors accordingly

* Revert "terminal: calc contrast and switch colors accordingly"

This reverts commit d0c30d5c4b34f605d0f6745546003f8ce7b0d9ec.

+++ lapce-app/src/config/terminal.rs
@@ -18,7 +18,7 @@ pub struct TerminalConfig {
     #[field_names(
         desc = "Set the terminal line height, If 0, it uses editor line height"
     )]
-    pub line_height: usize,
+    pub line_height: f64,
     #[field_names(desc = "Profiles available in terminal pane")]
     pub profiles: HashMap<String, TerminalProfile>,
     #[field_names(desc = "Default profile for each platform")]
